requirements.txt
azure-ai-vision-imageanalysis==1.0.0b2
azure-core==1.30.1
black==24.4.2
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
fonttools==4.52.1
frozendict==2.4.4
genanki==0.13.1
idna==3.7
imageio==2.34.1
imgaug==0.4.0
isodate==0.6.1
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.2
scikit-image==0.23.2
scipy==1.13.1
shapely==2.0.4
six==1.16.0
termcolor==2.4.0
tifffile==2024.5.22
typing_extensions==4.12.0
urllib3==2.2.1
